refactor(test): reuse watch clock fake
